Moved some functions over to package sessmngt in ./internal
