+++ .gitignore
@@ -3,3 +3,5 @@
 tmp/
 *.key
 *.pem
+main
+server
