+++ README.md
@@ -7,7 +7,7 @@ This is still a WIP, below is the features that will be implemented in the futur
 Urgent:
 * Either logout doesn't work or my page restriction is broken. Figure this out
 * Consolidation of all user/session management to its own go file
-* Redirection to user page after registration (Do after session management is consolidated)
+* Redirection to user page after registration or don't create a session until login (Do after session management is consolidated)
 * Server side validation of login sessions
 * Make navbar interactive to show if you are logged in, only display register button if not logged in
 * Link statistics for signed in users (Will need to figure out how to store on the backend)
