+++ go.mod
@@ -9,6 +9,7 @@ require (
 	gopkg.in/yaml.v2 v2.4.0
 )
 
+
 require (
 	github.com/dgrijalva/jwt-go v3.2.0+incompatible // indirect
 	github.com/golang-jwt/jwt v3.2.2+incompatible // indirect
